Copy implementation of thread_pool_base.
Inorder to use later the thread_pool_base we need to separate it from
tbb
